Progress indicator in the application icon
Displays a global progress indicator in the application icon reflecting the total progress of all running jobs.

Currently fully implemented on OSX (Cocoa).
On other OSes that do not allow to redraw the app icon, this can be implemented as a [x%] display in the app title, so to appear in the taskbar.

Thanks to Matt for the windowmanager wrapper.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@29143 954f8c5b-7b00-dc11-b283-0030488c597c