Avoid reset the H and S value in the color picker when V is equal to zero!
Small change to keep the value of H and S when V come to zero, take care
that only work if you keep the color picker open.

When the color picker is closed, the H and S value are reset to zero
this is because the color picker is used in a lot of different place
and this value need to be reset.

(BTW reset to zero only when V is equal to zero!)



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@29160 954f8c5b-7b00-dc11-b283-0030488c597c