== SoC Bullet - Bullet Upgrade to 2.76 ==
Updated Blender's Bullet to 2.76 in this branch only.

This update was done by:
1) deleting the contents of the existing extern/bullet2/src directory (leaving the .svn folder in place),
2) copy/pasting the contents of the bullet/src directory (from unzipped Bullet archive) into this newly cleared folder. 

Hopefully there aren't any patches that are still needed from the Bullet we had in source.

---

Note: I didn't use Moguri's patch, since that was giving me compile errors with headers not being able to be found.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@29506 954f8c5b-7b00-dc11-b283-0030488c597c