Bugfix #22030: Animation Editors and Layer Animation
Added new filtering option for animation editors (indicated as the 'ghost' toggle beside the 'select'(-ed only) toggle), which will include objects/bones that aren't visible (i.e. are hidden or on a hidden layer). 

This should make it possible to edit such types of animation, and also prevent flickering as these channels come in/out of view.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@29964 954f8c5b-7b00-dc11-b283-0030488c597c