Logic Editor UI: Make the Active object always the first one of the list, and have the ADD button only for it.
The reason for that is because I can't find a way to change the active object for a particular context (it may be even a bug in the rna/UI base code). So for the time being (a.k.a. for Beta) this will make it.

I actually like this solution, maybe the bug is for the good afterall.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@30256 954f8c5b-7b00-dc11-b283-0030488c597c