bugfix [#23656] Problems retrieving properties from inside runtime-created PointerProperties
cant test if this fix solves the problem matt is having but it at least fixes an error caused by classes created in exec(),
when the properties pointers were copied the hash key still referred to the python object which could be freed. in most cases this wouldnt happen (would be kept in bytecode) but with exec() the property string is freed immediately. 


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@31789 954f8c5b-7b00-dc11-b283-0030488c597c