Patch #24056: Indicate keyed status of swatches Submitted by: Shane Ambler (sambler)
<quote>
From blenderstorm idea# 169 this patch changes the swatch button to display a border showing the keyed status colour
of the colour.

I chose to go with showing the keyed colour in a border around the swatch, insetting the size of the swatch colour to
prevent altering the sizing of the swatch. An un-keyed swatch is not altered visually, only a keyed colour swatch is
changed to have the keyed colour around the outside of the swatch button.

I think a border of 3 pixels is sufficient to make the keyed colour visible without taking too much of the colour area
away.
</quote>

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@32214 954f8c5b-7b00-dc11-b283-0030488c597c