patch [#24178] bge.render.makeScreeshot - with help from Campbell(ideasman42) This patch brings back the old functionality from Blender 2.49. However we are forcing the format to be PNG only (as we had previously on blenderplayer).
Note: If letterboxing is on, we are recording only the camera area of the canvas (cool hein?).
Note2: I have a feeling that this is faster than what we had in 2.49 (which was really slow imo). Maybe it could be even faster if we disable PNG compression. Maybe an option for the future.

* patch finalized and committed as part of the BlenderPRO 2010 - BGE development workshop :) *

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@32387 954f8c5b-7b00-dc11-b283-0030488c597c