Partial fix for #22409: Locked axis + auto IK = bad
The "root" bone in an Auto-IK chain was never added properly if it didn't have a parent that it was connected to. This meant that if it had axis-locking (using transform locks), these would not get converted to temporary IK-locks. This also affects 2.49

The second part of the bug report though, is something more ingrained in the IK-solver internals (numeric error, which means that even locked axes aren't exactly untouched).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@32537 954f8c5b-7b00-dc11-b283-0030488c597c