Bugfix [#25298] Nasty group cycle that creates crash blend files is allowed
Infinite recursion (manisfesting as a crash) occurred when trying to
set the dupli-group setting on an object, when the object is a member
of the group being set. Added a check and warning for this in RNA to
prevent such setups from occurring in future.

Todo:
The warning report is currently only printed to console as I can't
quite remember how to grab reports pointer without context/operator
pointer available.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@33782 954f8c5b-7b00-dc11-b283-0030488c597c