Bugfix #25312
"Make Local" option "Objects and Data" made materials get
lost on save, the tagging was incorrect. Now it makes 
Materials local too.

Implementation note: the code is new (not 2.4) so it misses
some of the hairy goodies we had. It's something for another
time to really make this reliably work.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@33819 954f8c5b-7b00-dc11-b283-0030488c597c