Fix [#26474] double_sided wrong in COLLADA Export reported by Juan Linietsky
The export was not really wrong, but Maya expects their profile to be present inside <geometry>. Added this for mesh with ME_TWOSIDED flag set.

Ideally this will change in the future to be controllable through options to the exporter (like current Google Earth and 3DS Max extensions for <double_sided>).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@35713 954f8c5b-7b00-dc11-b283-0030488c597c