BGE Patch: [#27425] Allow to change the damping of the camera actuator ########## original name: "Allow to change the strenght of the "go behind" constraint of the camera actuator"
The camera actuator is an actuator that drive the camera to follow an object, with a set of constraint.
Currently, when the object followed rotate on himself (like a person, or an helicopter), the camera is really slow to go behind (at least 10 seconds).

This patch gives the UI to tweak the strenght of the 'go behind'[named damping] constraint.
###########

epydocs (rst) updated too

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@37455 954f8c5b-7b00-dc11-b283-0030488c597c