bugfix: [#28026] Copy Game Property broken
not exactly a bug, but the option to copy individual properties was not working from the SPACE menu.
I believe this was happening because we are using dynamic enums.

This commit makes the "merge" option to be the default one. So if you call it from the SPACE menu it will be the one used.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@38635 954f8c5b-7b00-dc11-b283-0030488c597c